refactor: rename extensions folder to util
It is really just a utilities folder
