+++ src/util/exceptions/ntta_error.h
