+++ src/util/random.h
