+++ src/util/string_extensions.h
